Chore(R9, R12, R14): organized the manager's archives
The top-clients component was generated too.
