+++ front-end/src/app/app.routes.ts
@@ -4,11 +4,12 @@ import { DepositComponent } from './features/deposit/deposit.component';
 import { UpdateProfileComponent } from './features/update-profile/update-profile.component';
 import { LoginComponent } from './features/login/login.component';
 import { RegisterComponent } from './features/register/register.component';
-import { ManagerHomeComponent } from './features/manager-home/manager-home.component';
+import { ManagerHomeComponent } from './features/manager/manager-home/manager-home.component';
 import {TransactionHistoryComponent} from './features/transaction-history/transaction-history.component';
 import {AdministratorHomeComponent} from './features/administrator/administrator-home/administrator-home.component';
 import {ListManagerComponent} from './features/administrator/list-manager/list-manager.component';
-import {ListaClientesComponent} from './features/client-list/client-list.component';
+import {ListaClientesComponent} from './features/manager/client-list/client-list.component';
+import { TopClientsComponent } from './features/manager/top-clients/top-clients.component';
 
 export const routes: Routes = [
 
@@ -22,6 +23,7 @@ export const routes: Routes = [
   { path: 'transaction-history', component: TransactionHistoryComponent },
   { path: 'administrator-home', component: AdministratorHomeComponent },
   { path: 'list-manager', component: ListManagerComponent },
-  {path: 'client-list',component:ListaClientesComponent}
+  {path: 'client-list',component:ListaClientesComponent},
+  {path: 'top-clients', component:TopClientsComponent}
 
 ];
