+++ front-end/src/app/features/manager/manager-home/components/deny-client/deny-client.component.css
