+++ front-end/src/app/features/manager/manager-home/manager-home.component.css
