+++ front-end/src/app/features/manager/top-clients/top-clients.component.html
@@ -1,0 +1,1 @@
+<p>top-clients works!</p>
