+++ front-end/src/app/features/manager/top-clients/top-clients.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-top-clients',
+  imports: [],
+  templateUrl: './top-clients.component.html',
+  styleUrl: './top-clients.component.css'
+})
+export class TopClientsComponent {
+
+}
